+++ lib/talis/hierarchy/asset.rb
@@ -54,7 +54,7 @@ module Talis
     #  asset.save
     class Asset < Talis::Resource
       extend Talis::OAuthService
-      include HierarchyResource
+      include Talis::Hierarchy::Resource
 
       base_uri Talis::BLUEPRINT_HOST
